file exist checking added!

+++ Analyzer.py
@@ -2,6 +2,7 @@ import tweepy
 import numpy as np
 import time
 from tqdm import tqdm
+import os
 
 from secrets import consumer_key, consumer_secret, access_token, access_token_secret
 auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
@@ -50,23 +51,31 @@ def get_friends(api, username, limit):
                 raise e
 
 def get_tweets(api, username, limit):
-    itr = tweepy.Cursor(api.user_timeline, screen_name=username).items(limit)
-    res = np.array([])
-    print( username )
+    if not os.path.isfile("./downloaded/" + username + ".db.ny"):
+        itr = tweepy.Cursor(api.user_timeline, screen_name=username).items(limit)
+        res = np.array([])
+        print( username )
 
-    while True:
-        try:
-            for status in tqdm(itr):
-                res = np.append(res, status)
-            np.save("./downloaded/"+username+".db", res)
-            return res
-        except tweepy.TweepError as e:
-            if e.reason == "Twitter error response: status code = 429":
-                print("Rate Limit !")
-                time.sleep(15 * 60)
-                print("Hi !")
-            else:
-                raise e
+        while True:
+            try:
+                for status in tqdm(itr):
+                    res = np.append(res, status)
+                np.save("./downloaded/"+username+".db", res)
+                return res
+            except tweepy.TweepError as e:
+                if e.reason == "Twitter error response: status code = 429":
+                    print("Rate Limit !")
+                    time.sleep(15 * 60)
+                    print("Hi !")
+                else:
+                    raise e
+    else:
+        print(username, " exist!")
 
 # data = getAllFavorites(api,usersList, 1000)
-get_friends(api, "sadeghhayeri", 5000)
+
+while True:
+    try:
+        get_friends(api, "sadeghhayeri", 5000)
+    except:
+        print("some errors! start over!")
